fix: sku

+++ src/order/services/order.service.ts
@@ -4,13 +4,15 @@ import {
   UnauthorizedException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/user/entities/user.entity';
 import { Repository } from 'typeorm';
 
 import { ProductRepository } from '../../product/repositories/product.repository';
-import { ProductService } from '../../product/services/product.service';
+import { ProductService } from '../../product/services/product.service';  
 import { Action } from '../../shared/acl/action.constant';
 import { Actor } from '../../shared/acl/actor.constant';
 import { RequestContext } from '../../shared/request-context/request-context.dto';
+import { UserService } from '../../user/services/user.service';
 import { CreateOrderDto } from '../dtos/create-order.dto';
 import { UpdateOrderDto } from '../dtos/update-order.dto';
 import { Order } from '../entities/order.entity';
@@ -27,18 +29,21 @@ export class OrderService {
     private productService: ProductService,
     private readonly ProductRepository: ProductRepository,
     private aclService: OrderAclService,
-    
+    private userService: UserService,
   ) {}
-
+  generateOrderCode(): string {
+    const timestamp = Date.now().toString();
+    const randomPart = Math.random().toString(36).substring(2, 8).toUpperCase();
+    return `ORD-${timestamp}-${randomPart}`;
+  }
   async create(
     createOrderDto: CreateOrderDto,
     ctx: RequestContext,
   ): Promise<Order> {
     const actor: Actor = ctx.user!;
-    console.log(actor);
     const isAllowed = this.aclService
       .forActor(actor)
-      .canDoAction(Action.Create, Order)
+      .canDoAction(Action.Create, Order);
 
     if (!isAllowed) {
       throw new UnauthorizedException();
@@ -47,7 +52,8 @@ export class OrderService {
     order.orderItems = [];
     order.userId = actor.id;
     let totalAmount = 0;
-
+    const orderCode = this.generateOrderCode();
+    order.orderCode = orderCode;
     for (const item of createOrderDto.orderItems) {
       const product = await this.productService.getProductById(
         ctx,
@@ -65,6 +71,11 @@ export class OrderService {
         { id: product.id },
         'stockQuantity',
         item.quantity,
+      );
+      await this.ProductRepository.increment(
+        { id: product.id },
+        'totalPurchases',
+        1,
       );
     }
 
@@ -89,16 +100,62 @@ export class OrderService {
     return order;
   }
 
-  async update(id: number, updateOrderDto: UpdateOrderDto): Promise<Order> {
+  async update(
+    ctx: RequestContext,
+    id: number,
+    updateOrderDto: UpdateOrderDto,
+  ): Promise<Order> {
+    const actor: Actor = ctx.user!;
+    console.log(actor);
+    const isAllowed = this.aclService
+      .forActor(actor)
+      .canDoAction(Action.Update, Order);
+
+    if (!isAllowed) {
+      throw new UnauthorizedException();
+    }
     const order = await this.findOne(id);
+    if (updateOrderDto.userId) {
+      const user = await this.userService.findById(ctx, updateOrderDto.userId);
+      if (!user) {
+        throw new NotFoundException(
+          `Người dùng với ID ${updateOrderDto.userId} không tìm thấy`,
+        );
+      }
+      order.user = user as unknown as User;
+    }
     Object.assign(order, updateOrderDto);
     return this.orderRepository.save(order);
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(ctx: RequestContext, id: number): Promise<void> {
+    const actor: Actor = ctx.user!;
+    const isAllowed = this.aclService
+      .forActor(actor)
+      .canDoAction(Action.Update, Order);
+
+    if (!isAllowed) {
+      throw new UnauthorizedException();
+    }
+    await this.orderItemRepository.delete({ order: { id: id } });
     const result = await this.orderRepository.delete(id);
     if (result.affected === 0) {
       throw new NotFoundException(`Order with ID ${id} not found`);
     }
+  }
+
+  async getOrdersByUser(ctx: RequestContext): Promise<Order[]> {
+    const actor: Actor = ctx.user!;
+    const isAllowed = this.aclService
+      .forActor(actor)
+      .canDoAction(Action.Read, Order);
+
+    if (!isAllowed) {
+      throw new UnauthorizedException();
+    }
+    return this.orderRepository.find({
+      where: { user: { id: actor.id } },
+      relations: ['orderItems', 'orderItems.product', 'user'],
+    });
   }
 }
